ODX-684 internal blueprint creations and default blueprint creation should not validate configs which can be setted by internal services.
